backend/requirements.txt
asgiref==3.9.1
astroid==3.3.11
autopep8==2.3.2
black==25.1.0
click==8.2.1
dill==0.4.0
Django==5.2.5
django-cors-headers==4.7.0
django-stubs==5.2.2
django-stubs-ext==5.2.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
flake8==7.3.0
isort==6.0.1
mccabe==0.7.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.3.8
pycodestyle==2.14.0
pyflakes==3.4.0
PyJWT==2.10.1
pylint==3.3.8
python-decouple==3.8
setuptools==80.9.0
sqlparse==0.5.3
tomlkit==0.13.3
types-PyYAML==6.0.12.20250809
types-requests==2.32.4.20250809
typing_extensions==4.14.1
urllib3==2.5.0
